requirements.txt
acres==0.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.31.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.5
langchain-core==0.3.13
langchain-text-splitters==0.3.1
langsmith==0.1.137
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msrest==0.7.1
multidict==6.1.0
narwhals==1.12.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.52.2
openpyxl==3.1.5
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.3
prov==2.0.1
puremagic==1.28
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.4
Pygments==2.18.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
scipy==1.15.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
yarl==1.17.0
